Use VC version to configure Windows builds
Instead of relying on Python to correctly configure Windows builds, use
the VC version from the matrix to configure Windows builds. Let this
determine the `vc` feature selected and tested against. Also adjust the
matrix via the `skip` to exclude the VC matrix on Unix and only use it
on Windows.