Skip building on VS pre-2015
Newer versions of NASM cannot be built on Visual Studio before 2015 or
more accurately it won't build on Visual Studio 2008. The constraint was
previously applied to this feedstock, but got lost in migration to
conda-build 3 and the new compiler syntax. This readds it. Hopefully
this fixes the Windows build issues.